+++ src/main/java/com/springboot/domain/member/service/MemberService.java
@@ -9,6 +9,8 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
+
 import static com.springboot.global.error.exception.ErrorCode.MEMBER_NOT_FOUND;
 
 @RequiredArgsConstructor
@@ -21,7 +23,7 @@ public class MemberService {
     }
 
     @Transactional
-    public MemberResponseDto get(Long memberId) {
+    public MemberResponseDto get(BigInteger memberId) {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new EntityNotFoundException(MEMBER_NOT_FOUND, "해당 id를 가진 member가 없습니다. id=" + memberId));
